Create the basic event-log database interface with a few events
Our design calls for an append-only database that tracks the interactions it
encounters, both for the elegance and peace-of-mind of never deleting anything,
and on the hunch that an event-stream based model would make it easier to merge
multiple databases in cases of concurrent editing.

The event-log presents an event-based low-level write interface, and a currently
extremely simple query interface that returns lists of constructed todo tasks.